Speed up OutlineShader on Intel GPUs
Some Intel integrated graphics do not cope well with the OutlineShader.
This can be observed most noticeably in the Outfitter and Shipyard where
it leads to a massive slowdown. This was discussed in #2442 and
addressed in #5179.

On my system with Intel UHD Graphics 620, the Linux 5.8 kernel's i915
driver, swaywm 1.5 with Xwayland, a USB-C-conntected 4K display and the
High-DPI plugin active I can observe a noticeable lag of mouse pointer
movements and ship switch via keyboard in the ShipInfo as well.

The comment above the fragment shader states: "To reduce sampling error
and bring out fine details, for every output pixel the Sobel filter is
actually applied in a 3x3 neighborhood and averaged together. That
neighborhood's scale is .618034 times the scale of the Sobel
neighborhood (i.e. the golden ratio) to minimize any aliasing effects
between the two."

Looking at the code, it appears that it actually offsets the center of
the Sobel filtering a .618034th of half a pixel derived from the sprite
size but then goes on to access texels at half-pixel offsets from
there. This access pattern seems to be very computationally expensive
for the Intel GPU. Removing the golden-ratio skew of the sampling center
from the shader removes all slowdown in Outfitter, Shipyard and
ShipInfo.

This change moves the golden ratio factor to the offset calculation,
yielding a Sobel filter neighborhood that is consistenly a .618034th of
half a pixel for all texture sampling accesses, removing the slowdown on
at least my system.